+++ app/src/main/java/com/udacity/gradle/builditbigger/ui/activity/MainActivity.java
@@ -7,7 +7,7 @@ import android.view.MenuItem;
 import android.view.View;
 import android.widget.Toast;
 
-import com.example.JokeTime;
+import com.joketimelib.JokeTime;
 import com.udacity.gradle.builditbigger.R;
